requirements.txt
Django==2.2
djangorestframework
drf_yasg
